auto_analyst/requirements.txt
streamlit
pandas
numpy
scipy
scikit-learn
mlxtend
plotly
matplotlib
seaborn
openpyxl
reportlab
jinja2
groq
python-dotenv
xlrd